[BUILD] Add rule to build map file with elf file together.
The generation of map file will be selectable by the user after this commit.

If the user adds dependency to .elf and .map (for example in the .images rule),
the new added rule will be automatically selected and map file will be generated.

When the user will have not the dependency on .map, the old rule will be selected
and no map file will be generated.

This should affect the seed of build process and give the user to have optionally
more debug info when it is needed.